requirements.txt
# Example requirements
pytest
langgraph
langchain_core
python-dotenv
python-dateutil